+++ frontend/components/screens/feed/Feed.module.scss
@@ -29,10 +29,6 @@
 					}
 					@apply mx-4 z-10;
 				}
-
-				.feed_container {
-					@apply pt-6 pb-8 flex flex-col gap-10 z-10;
-				}
 			}
 
 			// mobile-create-feed-btn
