+++ frontend/components/screens/feed/feed-container/FeedContainer.module.scss
@@ -1,0 +1,3 @@
+.feed_container {
+	@apply pt-6 pb-8 flex flex-col gap-10 z-10;
+}
