--- frontend/components/screens/feed/my-feed/MyFeed.module.scss
@@ -1,3 +1,0 @@
-.feed_container {
-	@apply pt-6 pb-8 flex flex-col gap-10;
-}
